chore(kiss): pair classification lives with its one producer
